#628 layout change
